mdadm: Remove klibc and uclibc support
Klibc is not working for at least 3 years and now and it has a conflict
with le/be_to_cpu() functions family provided by asm/byteorder.h which
will be added indirectly in next patches with raid/md_p.h.
Uclibc is not maintained and it is recomended to use musl instead.

It is right time to abandon those legacy targets. Remove them.

Signed-off-by: Mariusz Tkaczyk <mtkaczyk@kernel.org>