Rename is_name_posix_compatible to is_name_valid, allow :
The name of this function is a lie. The POSIX Portable Character
Set contains far more characters than just letters, digits and
_ . and -. Notably, it contains the space character, which this
function was specifically introduced to disallow.

The function also disallows the : character, which is a problem
because we often add this character to device names ourselves,
when we prepend the hostname. So with this check it is possible
to be blocked from assembling a device under the name mdadm
gave it:

mdadm --assemble /dev/md/system2:pv00 --run --uuid=8e4473bc:997208c1:da327078:0dc4d134 /dev/sda3 /dev/sdb1
mdadm: Value "system2:pv00" cannot be set as devname. Reason: Not POSIX compatible.

So, let's give this function an accurate name, and relax it to
allow the : character.

Signed-off-by: Adam Williamson <awilliam@redhat.com>