systemd: start mdcheck_continue.timer before mdcheck_start.timer
In the (unlikely but possible) case that a previously started md check
hasn't finished on the first Sunday of the following month,
mdcheck_start.service will start the scan from position 0, which is
probably not desired.

Have mdcheck_continue.service start first, so that it will pick up the
check where it left off, and that the subsequent mdcheck_start.service will
do nothing.

Signed-off-by: Martin Wilck <mwilck@suse.com>