mdcheck: restore backward-compatible behavior without arguments
The previous patch changed the behavior of mdcheck when invoked
without --continue or --restart. Previously it would have started
a new check from zero on all arrays. After the previous patch,
it acted like --continue on arrays where a check had been started already,
and started a new check only for arrays where no Checked_* marker
was present.

Introduce a new run mode --maybe-start for this behavior, and another
mode --force-start for the original behavior. --maybe-start is the
mode which will be called by mdcheck_continue.service now.

For backward compatibility reasons, --force-start is the default when no
mode argument (--continue, --maybe-start, --restart, or --force-start) is
specified. Because this is may interrupt checks that have already made
progress, the tool will print a warning and pause for a short time
when it's invoked in this manner.

Signed-off-by: Martin Wilck <mwilck@suse.com>