GitHub workflows: mdadm-test: new workflow for mdadm test suite
Since 3822896 ("github: disable self-runners"), the mdadm test suite
is generally disabled on GitHub. Introduce a new workflow which
executes on a standard GitHub-hosted runner.

Currently some tests fail. Fixing them is out of scope for this
patch. The main reason is that the kernel of the GitHub-hosted runners is
too old.

The workflow is configured such that it will only run when
dispatched manually, or when a push or PR against a branch named
'test_on_push' or 'test_on_pr', respectively.

Signed-off-by: Martin Wilck <mwilck@suse.com>